Changed file/module structure to improve clarity
