+++ src/lib.rs
@@ -73,8 +73,10 @@
 
 #![allow(clippy::needless_return)] // My preferred style
 
-mod address_modes;
 mod opcodes;
+mod address_modes;
+#[cfg(test)]
+mod test_utilities;
 
 #[macro_use]
 extern crate log;
@@ -348,45 +350,5 @@ enum StatusFlag {
 impl Default for MOS6502 {
     fn default() -> Self {
         MOS6502::new()
-    }
-}
-
-#[cfg(test)]
-pub(crate) struct StubInterface6502 {
-    read: fn(u16, u8) -> u8,
-    read_count: u8,
-    write: fn(u16, u8, u8),
-    write_count: u8,
-}
-
-#[cfg(test)]
-impl StubInterface6502 {
-    pub(crate) fn new(read_fn: fn(u16, u8) -> u8, write_fn: fn(u16, u8, u8)) -> Self {
-        StubInterface6502 { read: read_fn, write: write_fn, read_count: 0, write_count: 0 }
-    }
-}
-
-#[cfg(test)]
-impl Interface6502 for StubInterface6502 {
-    fn read(&mut self, address: u16) -> u8 {
-        self.read_count += 1;
-        (self.read)(address, self.read_count)
-    }
-
-    fn write(&mut self, address: u16, data: u8) {
-        self.write_count += 1;
-        (self.write)(address, data, self.read_count)
-    }
-}
-
-#[cfg(test)]
-impl Default for StubInterface6502{
-    fn default() -> Self {
-        StubInterface6502::new(|address, read_count| {
-            panic!("Read Function was not initialized")
-        },
-        |address, data, write_count| {
-            panic!("Write Function was not initialized")
-        })
     }
 }
